Rename ExtensionChatContext to ExtensionMessageContext
